const res = await mongoose.connect(process.env.MONGO_URI!);:
This line attempts to connect to the MongoDB database using the connection string stored in an environment variable MONGO_URI.
mongoose.connect: This method returns a promise which resolves to the Mongoose Connection object.
process.env.MONGO_URI!: The exclamation mark (!) is a TypeScript non-null assertion, indicating that MONGO_URI is expected to be defined.
isConnected = res.connection;:

If the connection is successful, isConnected is set to the Connection object from the result (res.connection).
console.log("Mongodb connected.");:

Logs a message indicating a successful connection to MongoDB.
return isConnected;:

Returns the Connection object, representing the established connection.