Interface Definitions export interface IPost {
Defines an IPost interface, outlining the structure of a post object.
description: string;

A string representing the description of the post.
user: IUser;

An object representing the user who created the post, adhering to the IUser interface.
imageUrl?: string;

An optional string representing the URL of the image associated with the post.
likes?: string[];

An optional array of strings representing the IDs of users who liked the post.
comments?: IComment[];

An optional array of comment objects adhering to the IComment interface.
export interface IPostDocument extends IPost, Document {

Defines an IPostDocument interface that extends both IPost and Document.
This interface represents a post document in MongoDB with additional Mongoose document properties.
createdAt: Date;

A date representing when the post was created.
updatedAt: Date;

A date representing when the post was last updated.