Interface Definitions Defines the IUser interface:
IUser: An interface that outlines the structure of a user object. It includes the following properties:
firstName: A string representing the user's first name.
lastName: A string representing the user's last name.
userId: A string representing a unique identifier for the user.
profilePhoto (optional): A string representing the URL of the user's profile photo.
bio (optional): A string representing the user's biography.