Schema Definition Defines a Mongoose schema named userSchema for IUserDocument:
firstName: A required string field for the user's first name.
lastName: A required string field for the user's last name.
userId: A required string field for the user's unique identifier.
profilePhoto: An optional string field for the URL of the user's profile photo, with a default value of an empty string.
bio: An optional string field for the user's biography, with a default value of an empty string.