JSX Structure <div> Wrapper:
The component is wrapped in a div element to provide a container for styling or additional structure.
<Avatar className='cursor-pointer'>:

The Avatar component is used to create a container for the profile photo.
The className='cursor-pointer' applies a CSS class to change the cursor to a pointer when hovering over the avatar, indicating it is clickable.
<AvatarImage src={src} alt="banner" />:

The AvatarImage component is used to display the profile photo.
src={src} sets the source of the image to the URL passed in the src prop.
alt="banner" provides alternative text for the image, which is useful for accessibility and in case the image cannot be loaded.