Database Connection Connects to the database: await connectDB(): This line ensures that the function establishes a connection to the MongoDB database before performing any database operations. Fetching the Post Finds the post by its ID:
const post = await Post.findById({_id: params.postId}): This line queries the database to find the post with the specified postId.
Handles case where post is not found:

if (!post) return NextResponse.json({error: "Post not found."}): If the post is not found in the database, the function returns a JSON response indicating that the post was not found.
Removing Like
Updates the post to remove the user's like:
await post.updateOne({$pull: {likes: userId}}): This line updates the post document to remove the userId from the likes array using MongoDB's $pull operator. This effectively removes the user's like from the post.
Returning Response
Returns success message:
return NextResponse.json({message: "Post disliked successfully."}): After successfully removing the like, the function returns a JSON response with a success message indicating that the post was disliked successfully.